v2/requirements.txt
streamlit
openai>=1.40.0
pydantic
python-docx
pdfplumber
